add middleware to set tenant and update flow, response messages
